+++ cpp/CHANGELOG.md
@@ -16,3 +16,4 @@ Please mark all change in change log and use the ticket from JIRA.
 
 - MS-1 - Add CHANGELOG.md
 - MS-4 - Refactor the vecwise_engine code structure
+- MS-6 - Implement SDK interface part 1